Move partner events to events package as 'external events'
