+++ website/events/admin/__init__.py
@@ -1,0 +1,3 @@
+from .event import *
+from .registration import *
+from .external_event import *
