+++ website/events/api/calendarjs/urls.py
@@ -4,6 +4,7 @@ from django.urls import path
 from events.api.calendarjs.views import (
     CalendarJSEventListView,
     CalendarJSUnpublishedEventListView,
+    CalendarJSExternalEventListView,
 )
 
 app_name = "events"
@@ -13,6 +14,11 @@ urlpatterns = [
         "events/",
         CalendarJSEventListView.as_view(),
         name="calendarjs-published",
+    ),
+    path(
+        "external/",
+        CalendarJSExternalEventListView.as_view(),
+        name="calendarjs-external",
     ),
     path(
         "events/unpublished/",
