+++ website/partners/api/v2/serializers/__init__.py
@@ -1,4 +1,3 @@
-from .partner_event import PartnerEventSerializer
 from .partner import PartnerSerializer
 from .vacancy import VacancySerializer
 from .vacancy_category import VacancyCategorySerializer
