+++ website/partners/models.py
@@ -216,58 +216,3 @@ class Vacancy(models.Model):
 
         ordering = ["-pk"]
         verbose_name_plural = _("Vacancies")
-
-
-class PartnerEvent(models.Model):
-    """Model describing partner event."""
-
-    partner = models.ForeignKey(
-        Partner,
-        verbose_name=_("partner"),
-        on_delete=models.CASCADE,
-        related_name="events",
-        blank=True,
-        null=True,
-    )
-
-    other_partner = models.CharField(max_length=255, blank=True)
-
-    title = models.CharField(_("title"), max_length=100)
-
-    description = models.TextField(_("description"))
-
-    location = models.CharField(
-        _("location"),
-        max_length=255,
-    )
-
-    start = models.DateTimeField(_("start time"))
-
-    end = models.DateTimeField(_("end time"))
-
-    url = models.URLField(_("website"))
-
-    published = models.BooleanField(_("published"), default=False)
-
-    def clean(self):
-        """Validate the partner event."""
-        super().clean()
-        errors = {}
-        if (not self.partner and not self.other_partner) or (
-            self.partner and self.other_partner
-        ):
-            errors.update(
-                {
-                    "partner": _("Please select or enter a partner for this event."),
-                    "other_partner": _(
-                        "Please select or enter a partner for this event."
-                    ),
-                }
-            )
-
-        if errors:
-            raise ValidationError(errors)
-
-    def __str__(self):
-        """Return the event title."""
-        return str(self.title)
