+++ website/thaliawebsite/api/calendarjs/urls.py
@@ -4,6 +4,5 @@ app_name = "thaliawebsite"
 
 urlpatterns = [
     path("", include("events.api.calendarjs.urls")),
-    path("", include("partners.api.calendarjs.urls")),
     path("", include("members.api.calendarjs.urls")),
 ]
